feat: extract to retrievals-api
